REFACTOR : Issue #125 Stage H2 Commit 3 - extract workflow/phase_a_common.rs [#125]
Extract the paradigm-agnostic post-tectonic-step pass into a new
default-features-on module so the C1 path (Commit 4) can call the
same implementation as the v2 path. Pure refactor — bit-identical
semantics preserved.

New module: workflow/phase_a_common.rs (default features, no
v2_legacy gate).

Public API:
  pub struct PostTectonicInput<'a> { ... }   // paradigm-agnostic
  pub struct PostTectonicOutput { common: CycleOutputCommon,
                                  new_cratonic_factor: Option<Field2D> }
  pub fn apply_post_tectonic(input) -> PostTectonicOutput
  pub fn extract_per_plate_type(plate_id, plate_type) -> Vec<PlateType>

Pipeline executed inside apply_post_tectonic (per user spec H2 §3):
  1. Sea-level threshold in S̃ space (Phase 3.5 fix preserved)
  2. Macro mass redistribution (Step 12 R3, mass-conserving)
  3. Reclassification (plate_type ← S̃ > sea_level_ref)
  4. Cratonic factor recompute (D4 retention rule)

phase_a_v2.rs rewrite:
- Drops the now-extracted helpers (extract_per_plate_type,
  recompute_cratonic_factor_for_cycle, measure_craton_change)
- Replaces inline Steps 2-6 with a single call to apply_post_tectonic
- Keeps the v2-specific Steps 1 (tectonic = run_baseline), 3
  (velocity reset, D1-ter empirical finding), and 4 (install
  cratonic_factor into FinalState — data-shape-specific)
- run_phase_a_cycle_v2 signature unchanged; CycleOutputV2 unchanged
- final_state_to_continuation_v2 unchanged

W2 audit: no v2-specific type leak into PostTectonicInput. Uses
only Field2D, PlateIdField, PlateTypeField, PlateType, PhaseAParams,
IsostasyConfig, CratonicConfigEnabled — all paradigm-agnostic.
tectonics_v2::cratonic is data-only and not v2_legacy-gated, so
CratonicConfigEnabled is accessible from default-features code.

W3 audit: shared scalars (mass_drift, erosion_volume_removed,
erosion_peak_delta_h, sea_level_normalized,
craton_recomputation_change) computed ONCE in apply_post_tectonic.
No duplication between paradigm-specific callers.

W4 audit: phase_a_common.rs has no cfg gate. Default-features-on
as designed.

Verification surface:
- cargo build --workspace --tests (default features) — clean
- cargo build -p ymir-core --tests --features v2_legacy — clean
- v2 path bit-identical regression:
  cargo test -p ymir-core --features v2_legacy
    --test v2_workflow_disabled_regression — 3/3 PASS including
    workflow_disabled_run_phase_a_cycle_is_bit_identical_to_run_baseline
- v2 path Enabled cratonic D4:
  cargo test -p ymir-core --features v2_legacy
    --test v2_workflow_phase_a_cycle — 3/3 PASS
- v2 macro mass conservation:
  cargo test -p ymir-core --test v2_workflow_macro — 2/2 PASS
- C1 path:
  cargo test -p ymir-core --test c1_phase_1_1_advection — 1/1 PASS
  cargo test -p ymir-core --test c1_phase_1_2_davis_suppe — 2/2 PASS
  cargo test -p ymir-core --lib tectonics_c1 — 42/42 PASS

Out-of-scope (pre-existing, W7 from Commit 1):
- cargo build -p ymir-viz --features v2_legacy still broken on
  Bevy ..default() / Transform — unrelated to this refactor.
